src/main/java/com/example/leetcode75/linkedlist/OddEvenLinkedList.java
package com.example.leetcode75.linkedlist;

public class OddEvenLinkedList {

    public static ListNode oddEvenList(ListNode head) {
        if (head == null || head.next == null || head.next.next == null) {
            return head;
        }

        ListNode even = head.next;
        ListNode odd = head;
        ListNode evenHead = even;
        ListNode oddHead = odd;

        while (even != null && even.next != null) {
            odd.next = even.next;
            odd = odd.next;
            even.next = odd.next;
            even = even.next;
        }
        odd.next = evenHead;
        return oddHead;
    }
}
